replace by a prisma update script
