+++ back/jest.config.integration.js
@@ -5,5 +5,6 @@ module.exports = {
     "^.+\\.(t|j)sx?$": ["@swc/jest"]
   },
   testEnvironment: "node",
-  testMatch: ["**/__tests__/**/?(*.)+(integration).[jt]s?(x)"]
+  testMatch: ["**/__tests__/**/?(*.)+(integration).[jt]s?(x)"],
+  testTimeout: 10000
 };
