+++ package.json
@@ -248,8 +248,6 @@
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
     "browserslist-useragent-regexp": "^2.0.5",
-    "commander": "^12.0.0",
-    "concurrently": "^6.4.0",
     "copyfiles": "^2.4.1",
     "core-js": "^3.21.1",
     "docusaurus-graphql-plugin": "^0.9.0",
@@ -288,7 +286,6 @@
     "typescript": "^5.0.2",
     "vite": "~4.3.9",
     "vite-plugin-svgr": "^4.0.0",
-    "ws": "^8.2.3",
-    "xlsx": "^0.18.5"
+    "ws": "^8.2.3"
   }
 }
